Dodanie nuke jako subkomenda komendy channel
